fix codacy analysis issue
